requirements.txt
appnope==0.1.3
asttokens==2.1.0
backcall==0.2.0
debugpy==1.6.3
decorator==5.1.1
distro==1.8.0
entrypoints==0.4
executing==1.2.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.0.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==2.11.2
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
six==1.16.0
stack-data==0.6.1
tabula-py==2.6.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
wcwidth==0.2.5
